2단계 - 볼링 점수판(그리기) (#300)
* refactor: 삭제 전 검증 메소드를 삭제 메소드 내에서 호출하도록 수정

1. 검증 메소드를 private 으로 변경
2. Answer.delete() 내에서 검증 메소드 호출
3. AnswerTest & AnswersTest 클래스 테스트 코드 수정

* feat: implement the input

1. implement controller & InputView
2. implement DTOs for input

* chore: delete empty.txt

* feat: implement entities

1. implement BowlingGame class
2. implement ScoreBoard class
3. implement Player class
4. implement Frame class

* refactor: make VO Pin class

* feat: implement & test Pin class

1. implement Pin class
2. test Pin class methods

* feat: implement & test Pitching class

1. implement Pitching class
2. test Pitching class methods

* feat: implement & test Pitchings class

1. implement NormalPitchings & LastPitchings class
2. test NormalPitchings & LastPitchings class methods

* feat: implement & test Frame & Frames class

1. implement Frame & Frames class
2. test Frame & Frames class methods

* feat: implement & test Player class

1. implement Player class
2. test Player class methods

* fix: fix typo

* fix: change test code in FramesTest

* refactor: add equals & hashcode methods in Player class

* feat: implement & test ScoreBoard class

1. implement ScoreBoard class
2. test ScoreBoard class methods

* feat: implement & test BowlingGame class

1. implement BowlingGame class
2. test BowlingGame class methods

* feat: implement DTO classes

* feat: implement controller & view classes

* fix: delete unused class FrameDTO

* refactor: f 문 스타일 가이드 참고하여 리팩토링

* refactor: 정규표현식 패턴을 정적 멤버 변수로 캐싱

* refactor: 정적 멤버 변수 상수 표기법으로 스타일 수정

* refactor: 다음 프레임을 계산해야 하는지 검증하는 로직을 메소드로 분리

* refactor: 정적 팩토리 메소드를 생성자 하단에 재배치

* refactor: if 문 스타일 가이드 참고하여 리팩토링

* refactor: 로직을 `getter` 로 꺼낸 데이터로 구현하는 대신, 메소드를 호출하도록 리팩토링

* refactor: 인터페이스 대신 제네릭 활용

* refactor(controller): add printScoreBoard() method

* refactor: 인스턴스 변수 3 개인 LastPitchings 클래스 리팩토링

+++ src/main/java/qna/domain/Answer.java
@@ -45,19 +45,20 @@ public class Answer extends AbstractEntity {
         this.contents = contents;
     }
 
-    public DeleteHistory delete() {
+    private void validateOwnerWhenDeleting(User loginUser) throws CannotDeleteException {
+        if (!writer.equals(loginUser)) {
+            throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
+        }
+    }
+
+    public DeleteHistory delete(User loginUser) throws CannotDeleteException {
+        validateOwnerWhenDeleting(loginUser);
         this.deleted = true;
         return new DeleteHistory(ContentType.ANSWER, getId(), writer, LocalDateTime.now());
     }
 
     public boolean isDeleted() {
         return deleted;
-    }
-
-    public void validateOwnerWhenDeleting(User loginUser) throws CannotDeleteException {
-        if (!writer.equals(loginUser)) {
-            throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
-        }
     }
 
     public User getWriter() {
